// src/protocol/enums.ts
export enum PayloadType {
	PROTO_MESSAGE = 5,
	ERROR_RES = 50,
	HEARTBEAT_EVENT = 51,

	APP_AUTH_REQ = 2100,
	APP_AUTH_RES = 2101,
	ACCOUNT_AUTH_REQ = 2102,
	ACCOUNT_AUTH_RES = 2103,
	VERSION_REQ = 2104,
	VERSION_RES = 2105,
	NEW_ORDER_REQ = 2106,
	TRAILING_SL_CHANGED_EVENT = 2107,
	CANCEL_ORDER_REQ = 2108,
	AMEND_ORDER_REQ = 2109,
	AMEND_POSITION_SLTP_REQ = 2110,
	CLOSE_POSITION_REQ = 2111,
	ASSET_LIST_REQ = 2112,
	ASSET_LIST_RES = 2113,
	SYMBOLS_LIST_REQ = 2114,
	SYMBOLS_LIST_RES = 2115,
	SYMBOL_BY_ID_REQ = 2116,
	SYMBOL_BY_ID_RES = 2117,
	SYMBOLS_FOR_CONVERSION_REQ = 2118,
	SYMBOLS_FOR_CONVERSION_RES = 2119,
	SYMBOL_CHANGED_EVENT = 2120,
	TRADER_REQ = 2121,
	TRADER_RES = 2122,
	TRADER_UPDATED_EVENT = 2123,
	RECONCILE_REQ = 2124,
	RECONCILE_RES = 2125,
	EXECUTION_EVENT = 2126,
	SUBSCRIBE_SPOTS_REQ = 2127,
	SUBSCRIBE_SPOTS_RES = 2128,
	UNSUBSCRIBE_SPOTS_REQ = 2129,
	UNSUBSCRIBE_SPOTS_RES = 2130,
	SPOT_EVENT = 2131,
	ORDER_ERROR_EVENT = 2132,
	DEAL_LIST_REQ = 2133,
	DEAL_LIST_RES = 2134,
	SUBSCRIBE_LIVE_TRENDBAR_REQ = 2135,
	UNSUBSCRIBE_LIVE_TRENDBAR_REQ = 2136,
	GET_TRENDBARS_REQ = 2137,
	GET_TRENDBARS_RES = 2138,
	EXPECTED_MARGIN_REQ = 2139,
	EXPECTED_MARGIN_RES = 2140,
	MARGIN_CHANGED_EVENT = 2141,
	OA_ERROR_RES = 2142,
	CASH_FLOW_HISTORY_REQ = 2143,
	CASH_FLOW_HISTORY_RES = 2144,
	GET_TICK_DATA_REQ = 2145,
	GET_TICK_DATA_RES = 2146,
	ACCOUNTS_TOKEN_INVALIDATED_EVENT = 2147,
	CLIENT_DISCONNECT_EVENT = 2148,
	GET_ACCOUNTS_BY_TOKEN_REQ = 2149,
	GET_ACCOUNTS_BY_TOKEN_RES = 2150,
	GET_CTID_PROFILE_REQ = 2151,
	GET_CTID_PROFILE_RES = 2152,
	ASSET_CLASS_LIST_REQ = 2153,
	ASSET_CLASS_LIST_RES = 2154,
	DEPTH_EVENT = 2155,
	SUBSCRIBE_DEPTH_REQ = 2156,
	SUBSCRIBE_DEPTH_RES = 2157,
	UNSUBSCRIBE_DEPTH_REQ = 2158,
	UNSUBSCRIBE_DEPTH_RES = 2159,
	SYMBOL_CATEGORY_LIST_REQ = 2160,
	SYMBOL_CATEGORY_LIST_RES = 2161,
	ACCOUNT_LOGOUT_REQ = 2162,
	ACCOUNT_LOGOUT_RES = 2163,
	ACCOUNT_DISCONNECT_EVENT = 2164,
	SUBSCRIBE_LIVE_TRENDBAR_RES = 2165,
	UNSUBSCRIBE_LIVE_TRENDBAR_RES = 2166,
	MARGIN_CALL_LIST_REQ = 2167,
	MARGIN_CALL_LIST_RES = 2168,
	MARGIN_CALL_UPDATE_REQ = 2169,
	MARGIN_CALL_UPDATE_RES = 2170,
	MARGIN_CALL_UPDATE_EVENT = 2171,
	MARGIN_CALL_TRIGGER_EVENT = 2172,
	REFRESH_TOKEN_REQ = 2173,
	REFRESH_TOKEN_RES = 2174,
	ORDER_LIST_REQ = 2175,
	ORDER_LIST_RES = 2176,
	GET_DYNAMIC_LEVERAGE_REQ = 2177,
	GET_DYNAMIC_LEVERAGE_RES = 2178,
	DEAL_LIST_BY_POSITION_REQ = 2179,
	DEAL_LIST_BY_POSITION_RES = 2180,
	ORDER_DETAILS_REQ = 2181,
	ORDER_DETAILS_RES = 2182,
	ORDER_LIST_BY_POSITION_REQ = 2183,
	ORDER_LIST_BY_POSITION_RES = 2184,
	DEAL_OFFSET_LIST_REQ = 2185,
	DEAL_OFFSET_LIST_RES = 2186,
	GET_POSITION_UNREALIZED_PNL_REQ = 2187,
	GET_POSITION_UNREALIZED_PNL_RES = 2188,
}